core: scripts: clean up and document core scripts
