+++ core/scripts/.gitignore
@@ -1,0 +1,2 @@
+.env
+debug-space-time-chart.html
