+++ core/scripts/README.md
@@ -8,3 +8,7 @@ They're meant to be used either directly as executable files (`./script.py`), or
 * `download-stdcm-requirements.py`: downloads the requirements for a given timetable, can be run to reproduce stdcm requests without loading the entire timetable
 * `download-timetable.py`: downloads a timetable that can be loaded in the scenario page
 * `generate-debug-space-chart.py`: generates a basic html with a space-time chart, using the json file generated by a successful stdcm request
+
+Some scripts can interact with production environments. This requires a gateway cookie.
+It can be obtained by visiting the web UI and looking for a `"gateway": "..."` cookie.
+It should then be placed in a .env file containing `GATEWAY_COOKIE="..."`
