+++ core/scripts/generate-debug-space-chart.py
@@ -21,7 +21,12 @@ from bokeh.models import DatetimeTickFormatter
 import click
 
 
-@click.command()
+@click.command(
+    help="""
+Reads the data generated by a successful stdcm simulation,
+and generates a basic HTML page containing a space-time chart of that new train.
+"""
+)
 @click.option(
     "--input-location",
     "-i",
